Removed user admin notification container
